+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.24] - 29.05.2025
+
+### Fixed
+- Custom target contract
+- Custom src folder handling
+
 ## [0.0.23] - 26.05.2025
 
 ### Fixed
@@ -37,6 +43,7 @@
 - Coverage report compatibility with new Medusa report format
 - Issue with fuzzer not stopping
 
+[0.0.24]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.24
 [0.0.23]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.23
 [0.0.22]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.22
 [0.0.21]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.21